+++ sensor_harm/landsat.py
@@ -27,13 +27,9 @@ LANDSAT_SCENE_PARSER = (
     r"(?P<path>[0-9]{3})"
     r"(?P<row>[0-9]{3})"
     r"_"
-    r"(?P<acquisitionYear>[0-9]{4})"
-    r"(?P<acquisitionMonth>[0-9]{2})"
-    r"(?P<acquisitionDay>[0-9]{2})"
+    r"(?P<sensing_date>[0-9]{8})"
     r"_"
-    r"(?P<processingYear>[0-9]{4})"
-    r"(?P<processingMonth>[0-9]{2})"
-    r"(?P<processingDay>[0-9]{2})"
+    r"(?P<processing_date>[0-9]{8})"
     r"_"
     r"(?P<collectionNumber>\w{2})"
     r"_"
@@ -41,8 +37,8 @@ LANDSAT_SCENE_PARSER = (
 )
 
 
-def get_landsat_angles(angle_dir: str, scene_id: str) -> Tuple[str, str, str, str]:
-    """Get Landsat angle bands file path.
+def landsat_angles(angle_dir: str, scene_id: str) -> Tuple[str, str, str, str]:
+    """Retrieve Landsat angle bands file path.
 
     Args:
         angle_dir (str): path to directory containing angle bands.
@@ -51,24 +47,35 @@ def get_landsat_angles(angle_dir: str, scene_id: str) -> Tuple[str, str, str, st
     """
     img_list = list(angle_dir.glob(f'**/{scene_id}*.tif'))
     logging.info('Load Landsat Angles')
-    pattern = re.compile('.*_solar_zenith_.*|.*_SZA.*')
-    sz_path = list(item for item in img_list if pattern.match(str(item)))[0]
-    pattern = re.compile('.*_solar_azimuth_.*|.*_SAA.*')
-    sa_path = list(item for item in img_list if pattern.match(str(item)))[0]
-    pattern = re.compile('.*_sensor_zenith_.*|.*_VZA.*')
-    vz_path = list(item for item in img_list if pattern.match(str(item)))[0]
-    pattern = re.compile('.*_sensor_azimuth_.*|.*_VAA.*')
-    va_path = list(item for item in img_list if pattern.match(str(item)))[0]
+    try:
+        pattern = re.compile('.*_solar_zenith_.*|.*_SZA.*')
+        sz_path = list(item for item in img_list if pattern.match(str(item)))[0]
+        pattern = re.compile('.*_solar_azimuth_.*|.*_SAA.*')
+        sa_path = list(item for item in img_list if pattern.match(str(item)))[0]
+        pattern = re.compile('.*_sensor_zenith_.*|.*_VZA.*')
+        vz_path = list(item for item in img_list if pattern.match(str(item)))[0]
+        pattern = re.compile('.*_sensor_azimuth_.*|.*_VAA.*')
+        va_path = list(item for item in img_list if pattern.match(str(item)))[0]
+    except:
+        raise RuntimeError(f'File not Found: Missing processed Angle bands on {angle_dir}')
 
     return sz_path, sa_path, vz_path, va_path
 
 
-def get_landsat_bands(satsen: str) -> Optional[List[str]]:
+def landsat_bands(parsed_sceneid: str) -> Optional[List[str]]:
     """Retrieve the bands which can be harmonized in Landsat data products."""
+    satsen = f'L{parsed_sceneid["sensor"]}{parsed_sceneid["satellite"]}'
+    collection = parsed_sceneid["collectionNumber"]
     if satsen == 'LT05' or satsen == 'LE07':
-        return ['sr_band1', 'sr_band2', 'sr_band3', 'sr_band4', 'sr_band5', 'sr_band7']
+        if collection=='01':
+            return ['sr_band1', 'sr_band2', 'sr_band3', 'sr_band4', 'sr_band5', 'sr_band7']
+        elif collection=='02':
+            return ['SR_B1', 'SR_B2', 'SR_B3', 'SR_B4', 'SR_B5', 'SR_B7']
     elif satsen == 'LC08':
-        return ['SR_B2', 'SR_B3', 'SR_B4', 'SR_B5', 'SR_B6', 'SR_B7']
+        if collection=='01':
+            return ['sr_band2', 'sr_band3', 'sr_band4', 'sr_band5', 'sr_band6', 'sr_band7']
+        elif collection=='02':
+            return ['SR_B2', 'SR_B3', 'SR_B4', 'SR_B5', 'SR_B6', 'SR_B7']
     return
 
 
@@ -89,18 +96,16 @@ def landsat_harmonize(scene_id: str, product_dir: str, target_dir: Optional[str]
         str: path to folder containing result images.
     """
     product_dir = Path(product_dir)
-    target_dir = Path(target_dir)
+    target_dir = Path(target_dir) / (scene_id + '_NBAR')
 
-    match = re.match(LANDSAT_SCENE_PARSER, scene_id, re.IGNORECASE)
+    parsed_sceneid = re.match(LANDSAT_SCENE_PARSER, scene_id, re.IGNORECASE)
 
-    if not match:
+    if not parsed_sceneid:
         raise RuntimeError(f'Invalid Landsat scene id {scene_id}')
-
-    satsen = f'L{match.group("sensor")}{match["satellite"]}'
 
     angle_dir = Path(angle_dir) if angle_dir else product_dir
     logging.info(f'Loading Angles from {angle_dir} ...')
-    sz_path, sa_path, vz_path, va_path = get_landsat_angles(angle_dir, scene_id)
+    sz_path, sa_path, vz_path, va_path = landsat_angles(angle_dir, scene_id)
 
     if target_dir is None:
         target_dir = product_dir.joinpath(Path('HARMONIZED_DATA'))
@@ -108,9 +113,9 @@ def landsat_harmonize(scene_id: str, product_dir: str, target_dir: Optional[str]
     target_dir.mkdir(parents=True, exist_ok=True)
 
     if bands is None:
-        bands = get_landsat_bands(satsen)
+        bands = landsat_bands(parsed_sceneid)
 
-    output_files = process_NBAR(product_dir, scene_id, bands, sz_path, sa_path, vz_path, va_path, satsen, target_dir)
+    output_files = process_NBAR(parsed_sceneid, product_dir, bands, sz_path, sa_path, vz_path, va_path, target_dir)
 
     # Copy quality band
     if cp_quality_band: